Refactor: Make component owner optional (#53)
When I initially created the ECS, I thought it would be a good idea to enforce "every component must be owned by some entity (e.g. a GameObject); it is invalid for a component not to have an owner." This is a pretty sensible idea, in theory - a component without an owner does nothing meaningful. However, it also makes constructing GameObjects hard at the moment and makes dependency injection tricky. We should be able to create components before a GameObject, and then pass them to the GameObject ctor - presently, we have to create the GameObject, then use its identity to construct Components, which we set with setter methods.

This PR makes it possible to create Components without EntityIDs, and updates GameObject ctor to set a component's ownerId.