Refactor: Remove component setters from GameObjects, and simplify GameObject loading in IntegrationWADLoader (#54)
Now that we can have components that don't need identities, we can just create them and pass them to the constructor of GameObject. We don't need setters; it's unlikely we'll need to reconstruct components after a GameObject is created, and if we need to we can probably just reconstruct the GameObject. (This probably should've been part of #53)