Refactor: Split realistic physics behavior into derived component  (#58)
This PR turns `PhysicsComponent` into an abstract base class, and implements `RealisticPhysicsComponent` and `SimplePhysicsComponent` as derived classes. We also introduce a few changes to the PhysicsManager so that certain realistic behaviors (i.e. gravity) are no longer part of it. **This PR breaks realistic collisions and gravity for the MVP**, neither of which we need to have working presently; porng will serve as the test app once completed.